+++ packages/prefabs/src/index.ts
@@ -12,3 +12,4 @@ export * from './traits.js';
 export * from './events.js';
 export * from './systems.js';
 export * from './commands.js';
+export * from './queries.js';